Unify the naming of the RPC calls
These are now in sync with the PHP implementation.
