+++ ts/WoltLabSuite/Core/Api/ModerationQueues/MarkAllModerationQueuesAsRead.ts
@@ -1,5 +1,5 @@
 /**
- * Marks a moderation queue as read.
+ * Marks all moderation queues as read.
  *
  * @author Olaf Braun
  * @copyright 2001-2025 WoltLab GmbH
@@ -11,20 +11,12 @@
 import { prepareRequest } from "WoltLabSuite/Core/Ajax/Backend";
 import { ApiResult, apiResultFromError, apiResultFromValue } from "../Result";
 
-type Response = {
-  unreadModerationItems: number;
-};
-
-export async function markAsRead(queueId: number): Promise<ApiResult<Response>> {
-  let response: Response;
-
+export async function markAllModerationQueuesAsRead(): Promise<ApiResult<[]>> {
   try {
-    response = (await prepareRequest(`${window.WSC_RPC_API_URL}core/moderation-queues/${queueId}/mark-as-read`)
-      .post()
-      .fetchAsJson()) as Response;
+    await prepareRequest(`${window.WSC_RPC_API_URL}core/moderation-queues/mark-all-as-read`).post().fetchAsJson();
   } catch (e) {
     return apiResultFromError(e);
   }
 
-  return apiResultFromValue(response);
+  return apiResultFromValue([]);
 }
