+++ ts/WoltLabSuite/Core/Api/ModerationQueues/MarkModerationQueueAsRead.ts
@@ -1,5 +1,5 @@
 /**
- * Marks all moderation queues as read.
+ * Marks a moderation queue as read.
  *
  * @author Olaf Braun
  * @copyright 2001-2025 WoltLab GmbH
@@ -11,12 +11,20 @@
 import { prepareRequest } from "WoltLabSuite/Core/Ajax/Backend";
 import { ApiResult, apiResultFromError, apiResultFromValue } from "../Result";
 
-export async function markAllAsRead(): Promise<ApiResult<[]>> {
+type Response = {
+  unreadModerationItems: number;
+};
+
+export async function markModerationQueueAsRead(queueId: number): Promise<ApiResult<Response>> {
+  let response: Response;
+
   try {
-    await prepareRequest(`${window.WSC_RPC_API_URL}core/moderation-queues/mark-all-as-read`).post().fetchAsJson();
+    response = (await prepareRequest(`${window.WSC_RPC_API_URL}core/moderation-queues/${queueId}/mark-as-read`)
+      .post()
+      .fetchAsJson()) as Response;
   } catch (e) {
     return apiResultFromError(e);
   }
 
-  return apiResultFromValue([]);
+  return apiResultFromValue(response);
 }
