+++ wcfsetup/install/files/js/WoltLabSuite/Core/Api/ModerationQueues/MarkAllModerationQueuesAsRead.js
